First version Launch Date
Now I'm not sure if this file should be in this repository.
It changes every time the release build script is run after all.
I decided to include it for the dead simple reason that people wanting to alter the launcher should not be forced to create a release build first.
Please note, this date will always be incorrect when you don't use my KthBuild.ps1 file to build (yeah, and it only works from powershell).