Empty function for scripting #68
This will be a very complex part of the editor, yet an important one, as I need this for precise customization. Star Story in particular wouldn't have been possible without this, but my other games are bound to this as well...

I am however planning to make this more expanded than before, plus more possibility to get area effects scriptable as well. This was (due to being in a hurry) not taken well in C#, but even the BMax version was a bit lacking here...

Also all scipts should be in Neil now, although, I may try to think of a "Lua Escape Route".