Fix Alpha255
This can also fix the issue that maps save in the C++ version of the Kthura Map Editor won't load properly in the C# editor and all programs using the C# library.
I must note, though, that I will NOT recompile the C# editor for this. It's deprecated and the C++ editor is already getting closer to the point where it can replace the C# editor entirely, so it won't make much sense at this point anymore.